modified:
  daemon/TODO
  daemon/algo/Algo/GatherHosts.t.cpp
  daemon/algo/Algo/countCorrelatedAlerts.t.cpp
  daemon/algo/Makefile@
  daemon/base/Base/NonEmptyVector.t.cpp
  daemon/base/Base/Threads/GrowingVector.t.cpp
  daemon/base/Base/Threads/detail/GrowingVectorData.t.cpp
  daemon/base/Base/ViaPointer.t.cpp
  daemon/base/Makefile@
  daemon/build_config/config.mk
  daemon/build_process/makefiles/common/tools.mk
  daemon/build_process/templates/build_config/config.mk
  daemon/commons/Makefile@
  daemon/configio/ConfigIO/FileReader.t.cpp
  daemon/configio/ConfigIO/Generic/Parse.t.cpp
  daemon/configio/ConfigIO/ParseFilters.t.cpp
  daemon/configio/ConfigIO/ParseInputs.t.cpp
  daemon/configio/ConfigIO/ParseLoggerAppenders.t.cpp
  daemon/configio/ConfigIO/ParsePersistency.t.cpp
  daemon/configio/ConfigIO/ParseTriggers.t.cpp
  daemon/configio/Makefile@
  daemon/core/Makefile@
  daemon/coretypes/Core/Types/Proc/Processor.t.cpp
  daemon/coretypes/Core/Types/Proc/ProcessorsCollection.t.cpp
  daemon/coretypes/Makefile@
  daemon/filter/Filter/BackendFacade.t.cpp
  daemon/filter/Filter/Factory.t.cpp
  daemon/filter/Filter/Simple/Strategy.t.cpp
  daemon/filter/Makefile@
  daemon/filterdnsresolver/Filter/DNSResolver/EntryProcessor.t.cpp
  daemon/filterdnsresolver/Filter/DNSResolver/Strategy.t.cpp
  daemon/filterdnsresolver/Makefile@
  daemon/filterhostcommon/Filter/HostCommon/Strategy.t.cpp
  daemon/filterhostcommon/Makefile@
  daemon/filteripblacklist/Filter/IPBlackList/EntryProcessor.t.cpp
  daemon/filteripblacklist/Filter/IPBlackList/Strategy.t.cpp
  daemon/filteripblacklist/Makefile@
  daemon/filtermanytomany/Filter/ManyToMany/Strategy.t.cpp
  daemon/filtermanytomany/Makefile@
  daemon/filtermanytoone/Makefile@
  daemon/filteronetomany/Makefile@
  daemon/filteronetoone/Filter/OneToOne/Strategy.t.cpp
  daemon/filteronetoone/Makefile@
  daemon/input/Input/Factory.t.cpp
  daemon/input/Input/Thread.t.cpp
  daemon/input/Makefile@
  daemon/inputfile/Makefile@
  daemon/inputprelude/Makefile@
  daemon/logger/Logger/Appenders/Base.hpp
  daemon/logger/Makefile@
  daemon/persistency/Makefile@
  daemon/persistency/Persistency/IO/Connection.t.cpp
  daemon/persistency/Persistency/NodeChildrenVector.t.cpp
  daemon/persistency/Persistency/detail/LimitedNULLString.t.cpp
  daemon/persistency/Persistency/detail/LimitedString.t.cpp
  daemon/persistencypostgres/Makefile@
  daemon/persistencypostgres/Persistency/IO/Postgres/Connection.t.cpp
  daemon/persistencypostgres/Persistency/IO/Postgres/StorageDataCache.t.cpp
  daemon/persistencypostgres/Persistency/IO/Postgres/detail/EntrySaver.t.cpp
  daemon/persistencystubs/Makefile@
  daemon/system/Makefile@
  daemon/system/System/AtExitImpl.t.cpp
  daemon/system/System/AtExitResourceDeallocator.t.cpp
  daemon/system/System/Backtrace.t.cpp
  daemon/testhelpers/Makefile@
  daemon/testhelperspersistency/Makefile@
  daemon/trigger/Makefile@
  daemon/trigger/Trigger/Factory.t.cpp
  daemon/xml/Makefile@
unknown:
  daemon/persistencypostgres/Persistency/IO/Postgres/Restorer.t.cpp.orig
pending merges:
  Bartek 'BaSz' Szu... 2010-04-07 switched from static libraries to dynamic (for faster linkage and smaller binar...
    Bartek 'BaSz' Szu... 2010-04-06 [merge] rebase.
    Bartek 'BaSz' Szu... 2010-04-06 build_process v1.5.0 (with distcc support) has been introduced.
    Bartek 'BaSz' Szu... 2010-04-06 TODO updated.
    Bartek 'BaSz' Szu... 2010-04-06 fixed signed/unsigned warnings from g++-4.1.
    Bartek 'BaSz' Szu... 2010-04-06 small fixes to prevent warnings from g++-4.1.
    Bartek 'BaSz' Szu... 2010-04-06 [merge] rebase.
    Bartek 'BaSz' Szu... 2010-04-06 [merge] rebase.
    Bartek 'BaSz' Szu... 2010-04-06 [merge] rebase.
    Bartek 'BaSz' Szu... 2010-03-31 [merge] rebase.

+++ daemon/input/Input/Thread.t.cpp
@@ -32,8 +32,8 @@ struct TestReader: public Reader
     return TestHelpers::Persistency::makeNewAlert();
   }
 
-  int  count_;
-  bool justThrow_;
+  size_t count_;
+  bool   justThrow_;
 }; // struct TestReader
 
 struct TestClass
@@ -72,8 +72,8 @@ template<>
 template<>
 void testObj::test<2>(void)
 {
-  ensure_equals("pre-condition 1 failed", tr_->count_, 0);
-  ensure_equals("pre-condition 2 failed", output_.size(), 0);
+  ensure_equals("pre-condition 1 failed", tr_->count_, 0u);
+  ensure_equals("pre-condition 2 failed", output_.size(), 0u);
   {
     boost::thread th( Thread(r_, output_) );
     // run ~2-3 times
@@ -82,7 +82,7 @@ void testObj::test<2>(void)
     th.interrupt();
     th.join();
   }
-  const int count=tr_->count_;
+  const size_t count=tr_->count_;
   usleep(100*1000);
   ensure_equals("thread didn't stopped", tr_->count_, count);
   ensure_equals("invalid queue size", output_.size(), count);
@@ -93,7 +93,7 @@ template<>
 template<>
 void testObj::test<3>(void)
 {
-  ensure_equals("pre-condition failed", tr_->count_, 0);
+  ensure_equals("pre-condition failed", tr_->count_, 0u);
   tr_->justThrow_=true;
   {
     boost::thread th( Thread(r_, output_) );
@@ -103,11 +103,11 @@ void testObj::test<3>(void)
     th.interrupt();
     th.join();
   }
-  const int count=tr_->count_;
+  const size_t count=tr_->count_;
   ensure("thread didn't run at all", count>0);
   usleep(100*1000);
   ensure_equals("thread didn't stopped", tr_->count_, count);
-  ensure_equals("invalid queue size", output_.size(), 0);
+  ensure_equals("invalid queue size", output_.size(), 0u);
 }
